Don't use the codacy action, use raw commands
The action uses a docker container and makes this mapping:
"/home/runner/work/nuclear-data-reader/nuclear-data-reader":"/github/workspace"

We then try and upload ../Build/coverage.info which becomes:
/github/workspace/../Build/coverage.info

This file doesn't exist. Is this because docker does not have it
mapped?

A build directory inside the repo may avoid this issue.