Use an enum to represent the measured state of an isotope
Previously we were using a uint8_t, but an enum is more descriptive.

We may have also uncovered, and squashed, a bug of not being
consistent with how the value was used between NUBASE and AME
data. Looks like we have got away with it as they are not compared or
used, but this may now have a knock on effect with the pynch project.