Upgrade the types of N and Z to avoid possible loss of data
We calculate the value of N by subtracting A from Z. However, A is
stored as uint16_t, but N and Z are uint8_t so the types did not
match. We know that the value of N will fit into a uint8_t but instead
of casting between values, lets just let N and Z have uint16_t types.