Compile into a static library on windows
For reasons I can't currently get to the bottom of, I'm not able to
make the necessary lib file which allows the library to be used. As a
result the tests can't link. The issue does not exist if building into
a static library so we will do that on windows for the moment.

Setting the path to the data directory was also causing issues
with compilation in one scenario and resolving the path if the
compilation was fixed. Setting the path via a definition flag allows
compilation and execution as expected.