Don't cast an enum class just so we can print it
You aren't meant to print the underlying type of an enum class, static
casting is a work around but lets not do that. This still has
hardcoded values, but I can't see them ever changing and I hope the
reason for their useage is clear. If it isn't, we can revisit at a
later date.