Generalise the removal of random spin parity substrings
If we store the substrings we are looking for in an array, we can use
a std::algorithm to check for them and then use a generic call to
erase. This removes the need for an if/else chain that may grow over
time. All we now need to do is add any new strings to the array and it
will automagically be taken care of.