Fix typos found by codespell
I've just learnt about the codespell package so lets use it. The clone
command also had a a typo which is now fixed.

At the moment I don't want to clutter the repository with yet another
file to allow us to ignore the element symbols during a codespell
check so we will ignore those warnings.