Enable converion warnings for GCC and clang
Past me didn't want to deal with fixing warnings related to comparing
the size of a container and the default int. Now that we are using
appropriate types, this is simpler to fix.

There are some static_casts which may be quick fixes so might need
revisiting in the future.