Revert 4d7836b so we don't have to ship the {fmt} header
Moving Converter::FloatToNdp to a template function created a dependency
on the {fmt} library with the converter.hpp header. We don't want that
so go back to defining the function in the cpp file. This may cause
issues with type conversion, but if it does, we'll revisit.

Running clang-format has indented the symbolZmap, I guess there was a
bug in previous versions that didn't do that? Looking at it, the new
behaviour appears to be correct.