Add a script that runs all the CMake presets
We want to test as much as possible before we push so this is a quality
of life script.

The common_functions script shouldn't enforce what is was enforcing, the
calling script can set up how it wants errors to be handled.

This new script outputs to log files which we don't want to track so
they are now ignored.